+++ examples/jit/film.py
@@ -6,6 +6,7 @@ from sklearn.metrics import f1_score
 from torch import Tensor
 from torch.nn import BatchNorm1d
 
+import torch_geometric.typing
 from torch_geometric.datasets import PPI
 from torch_geometric.loader import DataLoader
 from torch_geometric.nn import FiLMConv
@@ -47,7 +48,9 @@ class FiLM(torch.nn.Module):
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 model = FiLM(train_dataset.num_features, 320, train_dataset.num_classes,
              num_layers=4, dropout=0.1)
-model = torch.jit.script(model).to(device)
+if not torch_geometric.typing.WITH_PT212:
+    model = torch.jit.script(model)
+model = model.to(device)
 criterion = torch.nn.BCEWithLogitsLoss()
 optimizer = torch.optim.Adam(model.parameters(), lr=0.01)