+++ torch_geometric/nn/conv/graph_conv.py
@@ -1,9 +1,9 @@
 from typing import Final, Tuple, Union
 
-import torch
 from torch import Tensor
 
 from torch_geometric import EdgeIndex
+from torch_geometric._jit import is_scripting
 from torch_geometric.nn.conv import MessagePassing
 from torch_geometric.nn.dense.linear import Linear
 from torch_geometric.typing import Adj, OptPairTensor, OptTensor, Size
@@ -101,7 +101,7 @@ class GraphConv(MessagePassing):
         edge_weight: OptTensor,
     ) -> Tensor:
 
-        if not torch.jit.is_scripting() and isinstance(edge_index, EdgeIndex):
+        if not is_scripting() and isinstance(edge_index, EdgeIndex):
             return edge_index.matmul(
                 other=x[0],
                 input_value=edge_weight,