+++ torch_geometric/utils/mask.py
@@ -3,6 +3,7 @@ from typing import Optional
 import torch
 from torch import Tensor
 
+from torch_geometric._jit import is_scripting
 from torch_geometric.typing import TensorFrame
 
 
@@ -18,7 +19,7 @@ def mask_select(src: Tensor, dim: int, mask: Tensor) -> Tensor:
     """
     assert mask.dim() == 1
 
-    if not torch.jit.is_scripting():
+    if not is_scripting():
         if isinstance(src, TensorFrame):
             assert dim == 0 and src.num_rows == mask.numel()
             return src[mask]